# packages/api-client/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/sign-up": {
      "post": {
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpPayload"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignUpResult"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/sign-in": {
      "post": {
        "operationId": "AuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInPayload"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInResult"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/auth/profile": {
      "patch": {
        "operationId": "AuthController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfilePayload"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUser"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/assistant/internal/demographics/{userId}": {
      "get": {
        "operationId": "AssistantInternalController_getDemographics",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-llm-internal-secret",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Internal: fetch user demographics for tool calling",
        "tags": [
          "assistant-internal"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wira Borneo API",
    "description": "API documentation for local development and client generation",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "SignUpPayload": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "pregnantStatus": {
            "type": "object",
            "nullable": true
          },
          "ageGroup": {
            "type": "string",
            "nullable": true,
            "enum": [
              "UNDER_12",
              "AGE_12_17",
              "AGE_18_59",
              "AGE_60_PLUS"
            ]
          },
          "isPWD": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "AuthenticatedUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "image": {
            "type": "object",
            "nullable": true
          },
          "role": {
            "type": "object",
            "nullable": true
          },
          "pregnantStatus": {
            "type": "object",
            "nullable": true
          },
          "ageGroup": {
            "type": "string",
            "nullable": true,
            "enum": [
              "UNDER_12",
              "AGE_12_17",
              "AGE_18_59",
              "AGE_60_PLUS"
            ]
          },
          "isPWD": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "emailVerified"
        ]
      },
      "SignUpResult": {
        "type": "object",
        "properties": {
          "token": {
            "type": "object",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUser"
          }
        },
        "required": [
          "user"
        ]
      },
      "SignInPayload": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SignInResult": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUser"
          }
        },
        "required": [
          "token",
          "user"
        ]
      },
...
      "UpdateProfilePayload": {
        "type": "object",
        "properties": {
          "pregnantStatus": {
            "type": "object",
            "nullable": true
          },
          "ageGroup": {
            "type": "string",
            "nullable": true,
            "enum": [
              "UNDER_12",
              "AGE_12_17",
              "AGE_18_59",
              "AGE_60_PLUS"
            ]
          },
          "isPWD": {
            "type": "object",
            "nullable": true
          }
        }
      },
...
    }
  }
}